ai_slop_detector/requirements.txt
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
opencv-python-headless==4.8.1.78
imagehash==4.3.1
